:godmode: add user-scalable=no to viewport meta tag
The alternative to prevent browser zoom on input focus is to set the font-size of inputs to a minimum 16px. That is not feasible.
This meta tag change works okay for the most part as browsers ignore it when a user manually pinch and pan.
Although, pinch and pan to zoom is disabled in standalone PWA mode.